Account linking compability and removed user 388
While it is possible that you can link your account with multiple servers, you would be hitting the wall pretty quickly if you decide to reupload from a different server you linked your account with. Also I belive that nobody's aware that user 388 is used for reuploading, so I changed it to 0. This will (hopefully) not break anything and also thus not collide with other accounts, if someone else registered as user 388.